requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
opencv-python>=4.7.0
albumentations>=1.3.1
lime>=0.2.0.1
shap>=0.41.0
numpy>=1.23.5
pandas>=1.5.3
matplotlib>=3.7.1
streamlit>=1.22.0
pillow>=9.5.0
scikit-learn>=1.2.2 